Add files via upload
Also contains the U-Net model that is able to take the dataset generated by the script and train it. This also includes a pipeline at the end where it is possible to take an image input and output the mask at the end of the process using a trained model. Does include LaMa inpainting code directly taken(copied) from the LaMa colab example but, the functionality itself has not been implemented and it remains unmodified.