Fixed bugs in Mask_R_CNN_image_preprocessing.py where the files being copied and being preprocessed was not the mask but was the original file instead causing extremely long run times. Fixed it back to copying and processing Ground_Truth mask files.
Also added 'feature' in merge_groups_graph_reduce_WIP() function where it would automatically return from recursion when the algorithm fails to change the original input or if the total number of input groups and merged groups from the function is the same.
Quit recursion when this happens to stop any possible infinite recursions.